chore: implement FillLayer in TypeScript
