+++ docs/FillLayer.md
@@ -1,4 +1,4 @@
-<!-- This file was autogenerated from FillLayer.js do not modify -->
+<!-- This file was autogenerated from FillLayer.tsx do not modify -->
 # <MapboxGL.FillLayer />
 FillLayer is a style layer that renders one or more filled (and optionally stroked) polygons on the map.
 
@@ -11,10 +11,10 @@ FillLayer is a style layer that renders one or more filled (and optionally strok
 | aboveLayerID | `string` | `none` | `false` | Inserts a layer above aboveLayerID. |
 | belowLayerID | `string` | `none` | `false` | Inserts a layer below belowLayerID |
 | layerIndex | `number` | `none` | `false` | Inserts a layer at a specified index |
-| filter | `array` | `none` | `false` | Filter only the features in the source layer that satisfy a condition that you define |
+| filter | `FilterExpression` | `none` | `false` | Filter only the features in the source layer that satisfy a condition that you define |
 | minZoomLevel | `number` | `none` | `false` | The minimum zoom level at which the layer gets parsed and appears. |
 | maxZoomLevel | `number` | `none` | `false` | The maximum zoom level at which the layer gets parsed and appears. |
-| style | `union` | `none` | `false` | Customizable style attributes |
+| style | `FillLayerStyleProps` | `none` | `false` | Customizable style attributes |
 
 
 ## styles
